service/feature_store_warehouse: extract to separate service
This splist out the feature store logic that depends on a warehouse
into a separate service. This will keep the original feature store
service responsible for interacting with the mongo collection.

This will help keep the dependencies separate, and reduce the chance
for cyclical imports.
